site: merged html package into the site package site: ToHtml became TeiToHtml docbook: removed playwright
