+++ collector/src/main/scala/org/opentorah/collector/Collector.scala
@@ -1,8 +1,8 @@
 package org.opentorah.collector
 
-import org.opentorah.site.{Site, SiteCommon, SiteService}
+import org.opentorah.site.{LinksResolver, Site, SiteCommon, SiteService}
 import org.opentorah.store.{Alias, Context, Directory, ListFile, Path, Store, WithSource}
-import org.opentorah.tei.{EntityReference, EntityType, LinksResolver, Tei, Unclear, Entity as TeiEntity}
+import org.opentorah.tei.{EntityReference, EntityType, Tei, Unclear, Entity as TeiEntity}
 import org.opentorah.util.{Effects, Files}
 import org.opentorah.xml.{A, Caching, Element, Html, Parsable, Parser, ScalaXml, Unparser}
 import zio.{UIO, ZIO}
