+++ docbook/build.gradle
@@ -25,8 +25,6 @@ dependencies {
   implementation libs.saxon11
   implementation libs.bundles.fop
 
-  compileOnly libs.playwright
-
 //  implementation libs.graalvm.sdk
 
   // Needed to compile classes using Gradle, but does not become a dependency:
