+++ docs/_config.yml
@@ -7,24 +7,25 @@ description: >- # this means to ignore newlines until "baseurl:"
 baseurl: "" # the subpath of your site, e.g. /blog
 
 url: "http://www.opentorah.org" # the base hostname & protocol for your site, e.g. http://example.com
+
 timezone: America/New_York
+
 github_username: dubinsky
-# TODO add twitter...
+twitter_username: theopentorah
 
 # TODO Google Analytics?
 # google_analytics: UA-NNNNNNNN-N
 
-# TODO add back the tags for the blog?
-
 markdown: kramdown
 theme: minima
 
-#plugins:
+plugins:
 #  - jekyll-avatar
 #  - jekyll-feed
 #  - jekyll-sitemap
-# TODO add asciidoc
+- jekyll-asciidoc
 
+# TODO add back the tags for the blog?
 #header_pages:
 #  - tags.html
